Removed double implementation of vigenere_cipher.py. Fixes #8098
